Adding gradient accumulation and random cropping
Summary: This commit adds gradient accumulation strategy where
the gardients are accumulated for 'X' number of steps before
updating the weights, effectively mimicing the "iter_size"
functionality of Caffe. This allows for arbitrary length batch
sizes on the GPUs with less memory. Also, random cropping and mirroring
is added as means for better data augmentation.

Changes to be committed:
	modified:   README.md
	modified:   deeplab_resnet/image_reader.py
	modified:   evaluate.py
	modified:   fine_tune.py
	modified:   train.py
	modified:   train_msc.py

Signed-off-by: Arslan Chaudhry (arslan_mac@yahoo.com)